Rename constants to build config
