+++ tools/gulp/tasks/clean.ts
@@ -1,5 +1,5 @@
 import {task} from 'gulp';
-import {DIST_ROOT} from '../constants';
+import {DIST_ROOT} from '../build-config';
 import {cleanTask} from '../util/task_helpers';
