+++ tools/gulp/tasks/lint.ts
@@ -1,6 +1,6 @@
 import {task} from 'gulp';
 import {execNodeTask} from '../util/task_helpers';
-import {DIST_MATERIAL, DIST_CDK} from '../constants';
+import {DIST_MATERIAL, DIST_CDK} from '../build-config';
 
 /** Glob that matches all SCSS or CSS files that should be linted. */
 const stylesGlob = '+(tools|src)/**/*.+(css|scss)';
