+++ tools/gulp/tasks/payload.ts
@@ -1,7 +1,7 @@
 import {task} from 'gulp';
 import {join} from 'path';
 import {statSync} from 'fs';
-import {DIST_ROOT} from '../constants';
+import {DIST_ROOT} from '../build-config';
 import {spawnSync} from 'child_process';
 import {isTravisMasterBuild} from '../util/travis-ci';
 import {openFirebaseDashboardApp} from '../util/firebase';
